python/requirements.txt
psycopg2
confluent-kafka
sqlalchemy
python-dotenv
faker